move fill functions and util functions out of orc_chunk_reader.cpp
